move test component into mocks directory
